+++ info.rkt
@@ -1,6 +1,12 @@
 #lang info
+
+(define drracket-tools '(("tool.rkt")))
+(define drracket-tool-names '("sauron"))
+(define drracket-tool-icons '(#f))
+
 (define collection "sauron")
 (define deps '("base" "gui-lib" "net-lib"
+               "drracket" "drracket-plugin-lib"
                "drracket-tool-lib" "sandbox-lib"))
 (define build-deps '("scribble-lib" "racket-doc" "rackunit-lib"
                      "gui-doc"))